+++ src/main/java/insane96mcp/survivalreimagined/module/hungerhealth/feature/FoodDrinks.java
@@ -60,7 +60,7 @@ public class FoodDrinks extends SRFeature {
 	public static Boolean eatingSpeedBasedOffFood = true;
 	@Config
 	@Label(name = "Eating Speed Formula", description = "The formula to calculate the ticks required to eat a food. Variables as hunger, saturation_modifier, effectiveness as numbers and fast_food as boolean can be used. This is evaluated with EvalEx https://ezylang.github.io/EvalEx/concepts/parsing_evaluation.html. The default formula increases the time to eat exponentially when higher effectiveness, down to a minimum of 24 ticks (compared to 32 vanilla).")
-	public static String eatingSpeedFormula = "MAX((32 * effectiveness^1.4) / IF(fast_food, 2, 1) * 0.04, 24 / IF(fast_food, 2, 1))";
+	public static String eatingSpeedFormula = "MAX((32 * effectiveness^1.4) / IF(fast_food, 2, 1) * 0.04, 24 / IF(fast_food, 2, 1))"; //max((32 * x^1.4) * 0.04, 24)
 	@Config
 	@Label(name = "Stop consuming on hit", description = "If true, eating/drinking stops when the player's hit.")
 	public static Boolean stopConsumingOnHit = true;