+++ src/main/java/insane96mcp/survivalreimagined/mixin/EnchantmentMixin.java
@@ -12,13 +12,13 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(Enchantment.class)
 public class EnchantmentMixin {
 	@Inject(at = @At("RETURN"), method = "canEnchant", cancellable = true)
-	private void canApply(ItemStack stack, CallbackInfoReturnable<Boolean> callback) {
+	private void onCanEnchant(ItemStack stack, CallbackInfoReturnable<Boolean> callback) {
 		if (EnchantmentsFeature.disableEnchantment((Enchantment) (Object) this))
 			callback.setReturnValue(false);
 	}
 
 	@Inject(at = @At("RETURN"), method = "canApplyAtEnchantingTable", cancellable = true, remap = false)
-	private void canApplyAtEnchantingTable(ItemStack stack, CallbackInfoReturnable<Boolean> callback) {
+	private void onCanApplyAtEnchantingTable(ItemStack stack, CallbackInfoReturnable<Boolean> callback) {
 		if (EnchantmentsFeature.disableEnchantment((Enchantment) (Object) this))
 			callback.setReturnValue(false);
 	}