+++ src/main/java/insane96mcp/survivalreimagined/setup/SREnchantments.java
@@ -15,4 +15,5 @@ public class SREnchantments {
     public static final RegistryObject<Enchantment> BLASTING = REGISTRY.register("blasting", Blasting::new);
     public static final RegistryObject<Enchantment> EXPANDED = REGISTRY.register("expanded", Expanded::new);
     public static final RegistryObject<Enchantment> STEP_UP = REGISTRY.register("step_up", StepUp::new);
+    public static final RegistryObject<Enchantment> BANE_OF_SSSSS = REGISTRY.register("bane_of_sssss", BaneOfSSSS::new);
 }