+++ changelog.md
@@ -5,7 +5,7 @@
   * Fast but low durability
   * Iron tools are now made from Stone Tools instead of Flint
 * Adjusted the formula for healing from food
-  * Lower Hunger foods now real much less than higher hunger foods
+  * Lower Hunger foods now heal much less than higher hunger foods
 * Chickens can no longer be bred, use eggs
 
 ## Alpha 3.4.1